Moving supplemental-resources to it's own top-level project.
The project pom refers to the supplemental-resources module in the
Maven Remote Resources Plugin configuration.  This is essentially
a dependency: project -> supplemental-resources.

If you try to release the project pom, you'll have a problem. So
this commit makes supplemental-resources a direct child of
parent, just like project.  Now project can be released independently
of supplemental-resources.